refactoring groundwork
Laid some groundwork for refactoring, to remove the image processing from the main function to the imageManager. However, so far the work is not fully completed, it doesn't work as expected. I don't have time to follow this through, so I will (hopefully) come back to this at a later point.